+++ pysmt/simplifier.py
@@ -30,9 +30,7 @@ class Simplifier(walkers.DagWalker):
         self.functions[op.REAL_CONSTANT] = self.walk_identity
         self.functions[op.BOOL_CONSTANT] = self.walk_identity
         self.functions[op.INT_CONSTANT] = self.walk_identity
-
-        for bvop in op.BV_OPERATORS:
-            self.functions[bvop] = self.walk_identity
+        self.functions[op.BV_CONSTANT] = self.walk_identity
 
         self._validate_simplifications = None
 
@@ -392,5 +390,153 @@ class Simplifier(walkers.DagWalker):
             return self.manager.Real(args[0].constant_value())
         return self.manager.ToReal(args[0])
 
+    def walk_bv_and(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant():
+            lhs = args[0].constant_value()
+            rhs = args[1].constant_value()
+            res = lhs & rhs
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVAnd(*args)
 
- # EOC Simplifier
+    def walk_bv_not(self, formula, args):
+        if args[0].is_bv_constant():
+            res = ~args[0].constant_value() & (2**formula.bv_width() - 1)
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVNot(args[0])
+
+    def walk_bv_neg(self, formula, args):
+        if args[0].is_bv_constant():
+            res = 2**formula.bv_width() - args[0].constant_value()
+            res = res % 2**formula.bv_width()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVNeg(*args)
+
+    def walk_bv_or(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].constant_value() | args[1].constant_value()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVOr(*args)
+
+    def walk_bv_xor(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].constant_value() ^ args[1].constant_value()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVXor(*args)
+
+    def walk_bv_add(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].constant_value() + args[1].constant_value()
+            res = res % 2**formula.bv_width()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVAdd(*args)
+
+    def walk_bv_mul(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].constant_value() * args[1].constant_value()
+            res = res % 2**formula.bv_width()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVMul(*args)
+
+    def walk_bv_udiv(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            if args[1].bv_unsigned_value() == 0:
+                res = 2**formula.bv_width() - 1
+            else:
+                res = args[0].bv_unsigned_value() // args[1].bv_unsigned_value()
+                res = res % 2**formula.bv_width()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVUdiv(*args)
+
+    def walk_bv_urem(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            if args[1].bv_unsigned_value() == 0:
+                res = args[0].bv_unsigned_value()
+            else:
+                res = args[0].bv_unsigned_value() % args[1].bv_unsigned_value()
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVUrem(*args)
+
+    def walk_bv_ult(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].bv_unsigned_value() < args[1].bv_unsigned_value()
+            return self.manager.Bool(res)
+        return self.manager.BVULT(*args)
+
+    def walk_bv_ule(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].bv_unsigned_value() <= args[1].bv_unsigned_value()
+            return self.manager.Bool(res)
+        return self.manager.BVULE(*args)
+
+    def walk_bv_extract(self, formula, args):
+        if args[0].is_bv_constant():
+            bitstr = args[0].bv_bin_str(reverse=True)
+            start = formula.bv_extract_start()
+            end = formula.bv_extract_end()
+            res = bitstr[start:end+1][::-1]
+            w = (end+1)-start
+            return self.manager.BV("#b%s" % res, width=w)
+        return self.manager.BVExtract(args[0],
+                                      start=formula.bv_extract_start(),
+                                      end=formula.bv_extract_end())
+
+    def walk_bv_ror(self, formula, args):
+        if args[0].is_bv_constant():
+            bitstr = args[0].bv_bin_str(reverse=True)
+            # Takes first k elements and move to end
+            slice1 = bitstr[0:formula.bv_rotation_step()]
+            slice2 = bitstr[formula.bv_rotation_step():]
+            res = (slice2 + slice1)[::-1]
+            return self.manager.BV(res)
+        return self.manager.BVRor(args[0], formula.bv_rotation_step())
+
+    def walk_bv_rol(self, formula, args):
+        if args[0].is_bv_constant():
+            bitstr = args[0].bv_bin_str(reverse=True)
+            # Takes last k elements and move to beginning
+            slice1 = bitstr[0:-formula.bv_rotation_step()]
+            slice2 = bitstr[-formula.bv_rotation_step():]
+            res = (slice2 + slice1)[::-1]
+            return self.manager.BV(res)
+        return self.manager.BVRol(args[0], formula.bv_rotation_step())
+
+    def walk_bv_sext(self, formula, args):
+        if args[0].is_bv_constant():
+            bitstr = args[0].bv_bin_str()
+            filler = bitstr[0]
+            res = filler*formula.bv_extend_step() + bitstr
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVSext(args[0], formula.bv_extend_step())
+
+    def walk_bv_zext(self, formula, args):
+        if args[0].is_bv_constant():
+            bitstr = args[0].bv_bin_str()
+            filler = "0"
+            res = filler*formula.bv_extend_step() + bitstr
+            return self.manager.BV(res, width=formula.bv_width())
+        return self.manager.BVZExt(args[0], formula.bv_extend_step())
+
+    def walk_bv_concat(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            w0 = args[0].bv_width()
+            w1 = args[1].bv_width()
+            res = (2**w1) * args[0].bv_unsigned_value() + \
+                  args[1].bv_unsigned_value()
+            return self.manager.BV(res, w1 + w0)
+        return self.manager.BVConcat(*args)
+
+    def walk_bv_lshl(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].bv_unsigned_value() << args[1].bv_unsigned_value()
+            w = args[0].bv_width()
+            return self.manager.BV(res % (2 ** w), w)
+        return self.manager.BVLShl(*args)
+
+    def walk_bv_lshr(self, formula, args):
+        if args[0].is_bv_constant() and args[1].is_bv_constant:
+            res = args[0].bv_unsigned_value() >> args[1].bv_unsigned_value()
+            w = args[0].bv_width()
+            return self.manager.BV(res % (2 ** w), w)
+        return self.manager.BVLShr(*args)
+
+# EOC Simplifier